requirements.txt
streamlit
faiss-cpu
sentence-transformers
zhipuai
pandas
numpy